credentials_manager.py
This is the second version of this tool. New features : -No duplicated accounts can be in the same platform.                         -You can now get only one account credentials and of course you can get them all too. Check the --help or -h flags to see more.                         -No one can access or see what inside your credentials file except you (admin in windows and root (sudo) in linux).                         -The exact date and time is displayed for every credential so you can know when you have changed it or implemented it.                   In the future i'll make it more reliable so stay tuned Enjoy...